requirements.txt
flask==2.2.3
flask-bootstrap==3.3.7.1
flask-ckeditor==0.4.6
flask-login==0.6.2
flask-sqlalchemy==3.0.3
flask-wtf==1.1.1
flask-mail==0.9.1
werkzeug==2.2.3
wtforms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
flask-migrate==4.0.4
# Additional packages for deployment
gunicorn==21.2.0
psycopg2-binary==2.9.9